omni: Update backuptool(_ab) to ADDONDD_VERSION V3
backuptool_ab: Make copy_file preserve file/directory attrs


prebuilt: backuptool: Implement dynamic system mount for A-only

* Dynamically mount system to the path chosen by the recovery through backuptool
* This can be helpful because of the fragmentation that will happen with system mount in recovery after Q


backuptool: Fix addon.d version check for A-only

 * The path /postinstall exists only for A/B, causing:
   grep: /postinstall/tmp/addon.d/*sh: No such file or directory


backuptool: Execute check_blacklist from the current directory

 * For some odd reasons executing `cd /system/addon.d` makes the system
   hang and unmount error:
     umount: /system_root: Device or resource busy
 * Don't change directory to not allow the system partition hang


prebuilt: backuptool: Drop addond version check on A-only

 * This check was supposed to check whether the script
   addon.d version was lower than backuptool's

 * Given that the backuptool addon.d version is 1, this
   isn't going to happen ever making this check completely
   unuseful


backuptool: Remove backup blacklist


backuptool: Properly unmount system partition

For non AB devices system partition should be unmounted
if check_prereq function fails.

This patch also refactors backuptool a bit for AB devices
in order to look same as backuptool for non AB devices.


backuptool: Fix backup/restore functionality

Backup/restore functionality was broken in the
Ia1f4ae95c9e4dae4df844853e81c264bc838f177 change
because of incorrect check of the function's result.

check_prereq() function refactored to return 0 if
backuping/restoration is possible. Any work should be
performed only if check_prereq() succeeds.


backuptool_ab: Use mv instead of cp for restore step

Using cp instead of mv during the restore step resulted in double the
amount of free space and inodes being used.
For example, a nano arm64 opengapps install uses around 500M. But during
a backup/restore process when the Updater is used, it temporarily uses
1G on the other slot.  That causes failed updates on devices that
don't have that much free space on the system partition.


backuptool_ab: Make tmp actually tmpfs


backuptool: Preserve symlinks

Without -d, we copy the file sysmlinked rather than the symlink.


backuptool: Preserve the SELinux context of the files

Add a function that allows to copy files preserving their SELinux
context that is generic enough to work with both busybox and toybox.


backuptool: Support seamless backup and restore to extra partitions

For scripts declaring ADDOND_VERSION=3 automatically mount
vendor, product, system_ext and others (when they're dedicated partitions).

Also expose the get_output_path() function to get the path to where
a file is mounted in case it lives in a dedicated partition.

ab exapmles:
get_output_path "system/product/priv-app/MyApp.apk"  = "/postinstall/product/priv-app/MyApk.apk"
get_output_path "system/app/MySystemApp.apk"         = "/postinstall/system/app/MySystemApp.apk"

a-only examples:
get_output_path "/mnt/system/system/product/priv-app/MyApp.apk" = "/mnt/system/system/product/priv-app/MyApp.apk"

******************************************************************
Instead of cycling all scripts for each stage, run
pre-backup -> backup -> post-backup in quick succession
(and likewise for restore), to ensure backwards compatibility
with scripts that wrongly assumed their environment not to
change between steps.
This is needed because we want to undo any mounting done for V3
scripts when executing V2 scripts. If a V2 script did mounting in
pre-restore and expected things to still be mounted in restore,
we would break their (yes incorrect) assumption.


backuptool: Correct supported partitions for addon.d v3

* The plan was to support odm & oem too, but it turned out
  unnecessarily complicated due to their symlink setup,
  while being most likely unused.
  The partitions were removed from the list of supported ones
  in a-only backuptool, but were forgotten in ab backuptool.


backuptool: Ensure to cleanup the environment on failure


backuptool: Do not set ADDOND_VERSION=2 for a-only

There are addon.d scripts that rely on the value of ADDOND_VERSION
to determine if they're being called from a-only vs a/b backuptool.

If they declare ADDOND_VERSION=3, they shall stop doing that;
otherwise offer them the same environment, that is unset ADDOND_VERSION
for a-only backuptool.


backuptool: Ensure dedicated partitions are not unused before mounting

* Ignore the block devices in case their mount points are symlinks.
  This is common on devices where maintainers have chosen not to use
  real partitions because of their size being too small to be useful

Also `continue` instead of `break`. Oops.


backuptool: use correct BLK_PATH on a-only dynamic partitions devices


backuptool: Support tab-delimited fstab


backuptool: Unmount v3 partitions during backup

Extracted file /tmp/install/bin/backuptool.functions
Extracted file /tmp/install/bin/backuptool.sh
Extracted 2 file(s)
about to run program [/tmp/install/bin/backuptool.sh] with 5 args
[libfs_mgr]Unmapped logical partition system
DM_DEV_REMOVE failed for [vendor]: Device or resource busy
Cannot unmap vendor before removing group qti_dynamic_partitions.
script aborted: assert failed: update_dynamic_partitions(package_extract_file(dynamic_partitions_op_list))
assert failed: update_dynamic_partitions(package_extract_file(dynamic_partitions_op_list))error: 25
Updater process ended with ERROR: 1
Failed to mount '/system_root' (No such file or directory)
I:Actual block device: '/dev/block/dm-0', current file system: 'ext4'

We can clearly see that after version 3 script is executed with
its specific functions. It cannot or don't unmount partitions for
successful OTA upgrade. Resulting broken installation, this also
affects ROM inbuilt OTA updater app.

+++ prebuilt/common/bin/backuptool.sh
@@ -4,28 +4,28 @@
 #
 
 export C=/tmp/backupdir
-export S=$2
+export SYSDEV="$(readlink -nf "$2")"
+export SYSFS="$3"
 export V=13
 
-export ADDOND_VERSION=1
-
 DEBUG=0
+
+export ADDOND_VERSION=3
+
+# Partitions to mount for backup/restore in V3
+export all_V3_partitions="vendor product system_ext"
+
+get_script_version() {
+  version=$(grep "^# ADDOND_VERSION=" $1 | cut -d= -f2)
+  [ -z "$version" ] && version=1
+  echo $version
+}
 
 # Preserve /system/addon.d in /tmp/addon.d
 preserve_addon_d() {
   rm -rf /tmp/addon.d/
   mkdir -p /tmp/addon.d/
   cp -a $S/addon.d/* /tmp/addon.d/
-    # Discard any scripts that aren't at least our version level
-    for f in /postinstall/tmp/addon.d/*sh; do
-      SCRIPT_VERSION=$(grep "^# ADDOND_VERSION=" $f | cut -d= -f2)
-      if [ -z "$SCRIPT_VERSION" ]; then
-        SCRIPT_VERSION=1
-      fi
-      if [ $SCRIPT_VERSION -lt $ADDOND_VERSION ]; then
-        rm $f
-      fi
-    done
   chmod 755 /tmp/addon.d/*.sh
 }
 
@@ -37,35 +37,126 @@ restore_addon_d() {
 
 # Restore only if backup has the expected major and minor version
 check_prereq() {
-  if [ ! -f /tmp/build.prop ]; then
-    # this will block any backups made before 8 cause file was not copied before
-    echo "Not restoring files from incompatible version: $V"
-    exit 127
-  fi
+  # If there is no build.prop file the partition is probably empty.
+if [ ! -r $S/build.prop ]; then
+  echo "Backup/restore is not possible. Partition is probably empty"
+  return 1
+fi
   if ( ! grep -q "^ro.build.version.release=$V.*" /tmp/build.prop ); then
-    echo "Not restoring files from incompatible version: $V"
-    exit 127
+    echo "Backup/restore is not possible. Incompatible ROM version: $V"
+  return 2
   fi
+  return 0
 }
 
-check_blacklist() {
-  if [ -f $S/addon.d/blacklist ];then
-    ## Discard any known bad backup scripts
-    cd $1/addon.d/
-    for f in *sh; do
-      s=$(md5sum $f | awk {'print $1'})
-      grep -q $s $S/addon.d/blacklist && rm -f $f
-    done
-  fi
-}
-
-# Execute /system/addon.d/*.sh scripts with $1 parameter
-run_stage() {
+# Execute /system/addon.d/*.sh scripts with each $@ parameter
+run_stages() {
   for script in $(find /tmp/addon.d/ -name '*.sh' |sort -n); do
     if [ $DEBUG -eq 1 ]; then
         echo run_stage $script $1
     fi
-    $script $1
+    v=$(get_script_version $script)
+    if [ $v -ge 3 ]; then
+      mount_extra $all_V3_partitions
+    else
+      umount_extra $all_V3_partitions
+    fi
+
+    for stage in $@; do
+      if [ $v -ge 3 ]; then
+        $script $stage
+      else
+        ADDOND_VERSION= $script $stage
+      fi
+    done
+  done
+}
+
+#####################
+### Mount helpers ###
+#####################
+determine_system_mount() {
+  if grep -q -e"^$SYSDEV" /proc/mounts; then
+    umount $(grep -e"^$SYSDEV" /proc/mounts | cut -d" " -f2)
+  fi
+
+  if [ -d /mnt/system ]; then
+    SYSMOUNT="/mnt/system"
+  elif [ -d /system_root ]; then
+    SYSMOUNT="/system_root"
+  else
+    SYSMOUNT="/system"
+  fi
+
+  export S=$SYSMOUNT/system
+}
+
+mount_system() {
+  mount -t $SYSFS $SYSDEV $SYSMOUNT -o rw,discard
+}
+
+unmount_system() {
+  umount $SYSMOUNT
+}
+
+get_block_for_mount_point() {
+  grep -v "^#" /etc/recovery.fstab | grep "[[:blank:]]$1[[:blank:]]" | tail -n1 | tr -s [:blank:] ' ' | cut -d' ' -f1
+}
+
+find_block() {
+  local name="$1"
+  local fstab_entry=$(get_block_for_mount_point "/$name")
+  # P-SAR hacks
+  [ -z "$fstab_entry" ] && [ "$name" = "system" ] && fstab_entry=$(get_block_for_mount_point "/")
+  [ -z "$fstab_entry" ] && [ "$name" = "system" ] && fstab_entry=$(get_block_for_mount_point "/system_root")
+
+  local dev
+  if [ "$DYNAMIC_PARTITIONS" = "true" ]; then
+    if [ -n "$fstab_entry" ]; then
+      dev="${BLK_PATH}/${fstab_entry}"
+    else
+      dev="${BLK_PATH}/${name}"
+    fi
+  else
+    if [ -n "$fstab_entry" ]; then
+      dev="$fstab_entry"
+    else
+      dev="${BLK_PATH}/${name}"
+    fi
+  fi
+
+  if [ -b "$dev" ]; then
+    echo "$dev"
+  fi
+}
+
+determine_system_mount
+
+DYNAMIC_PARTITIONS=$(getprop ro.boot.dynamic_partitions)
+if [ "$DYNAMIC_PARTITIONS" = "true" ]; then
+    BLK_PATH="/dev/block/mapper"
+else
+    BLK_PATH=$(dirname "$SYSDEV")
+fi
+
+mount_extra() {
+  for partition in $@; do
+    mnt_point="/$partition"
+    mountpoint "$mnt_point" >/dev/null 2>&1 && continue
+    [ -L "$SYSMOUNT/$partition" ] && continue
+
+    blk_dev=$(find_block "$partition")
+    if [ -e "$blk_dev" ]; then
+      [ "$DYNAMIC_PARTITIONS" = "true" ] && blockdev --setrw "$blk_dev"
+      mkdir -p "$mnt_point"
+      mount -o rw "$blk_dev" "$mnt_point"
+    fi
+  done
+}
+
+umount_extra() {
+  for partition in $@; do
+    umount -l "/$partition" 2>/dev/null
   done
 }
 
@@ -75,24 +166,26 @@ case "$1" in
     rm -rf $C
     cp $S/bin/backuptool.functions /tmp
     cp $S/build.prop /tmp
-    mkdir -p $C
-    #check_prereq
-    check_blacklist $S
-    preserve_addon_d
-    run_stage pre-backup
-    run_stage backup
-    run_stage post-backup
+    mount_system
+    if check_prereq; then
+      mkdir -p $C
+      preserve_addon_d
+      run_stages pre-backup backup post-backup
+      umount_extra $all_V3_partitions
+    fi
+    unmount_system
   ;;
   restore)
     cp $S/bin/backuptool.functions /tmp
-    check_prereq
-    check_blacklist /tmp
-    run_stage pre-restore
-    run_stage restore
-    run_stage post-restore
-    restore_addon_d
-    rm -rf $C
-    sync
+    mount_system
+    if check_prereq; then
+      run_stages pre-restore restore post-restore
+      umount_extra $all_V3_partitions
+      restore_addon_d
+      rm -rf $C
+      sync
+    fi
+    unmount_system
   ;;
   *)
     echo "Usage: $0 {backup|restore}"
